Refactor: MainSection Logic 분리
