+++ src/pages/TypingPage/MainSection/index.css
@@ -26,6 +26,7 @@
   display: flex;
   flex-direction: column;
   align-items: center;
+  row-gap: 20px;
 }
 
 .statistics_button {
